event skeleton

+++ backend/Entities/Event.cs
@@ -2,11 +2,11 @@
 
 public class Event
 {
-    public long Start { get; set; }
-    public long End { get; set; }
+    public int Start { get; set; }
+    public int End { get; set; }
     public string Description { get; set; }
 
-    public Event(long start, long end, string description)
+    public Event(int start, int end, string description)
     {
         Start = start;
         End = end;
